src/backend/requirements.txt
Django==1.9.7
gevent==1.2.1
grequests==0.3.0
requests==2.5.0
Pillow==4.0.0
beautifulsoup4==4.5.3
